feat(icons-list): make icons list page optional
This commit adds a module option called `buildIconsList` that triggers
the creation of the icons list page in development mode, previously
added in 4425e7c54d ("feat(icons-list): list all icons in a single page
on dev mode (#108)").

This effectively makes possible to disable the creation of this test
page in development mode; although it has been kept to true as default,
so it will be built unless told otherwise in nuxt.config.js.